requirements.txt
torch>=1.9.0
torchvision>=0.10.0
timm>=0.4.12
Pillow>=8.0.0
matplotlib>=3.3.0
numpy>=1.19.0
Flask>=2.0.0
flask-cors>=3.0.10
pandas>=1.3.0
scikit-learn>=1.0.0
opencv-python>=4.5.0
requests>=2.25.0
tqdm>=4.60.0
